+++ tasti-labs/render-details.js
@@ -1,0 +1,5 @@
+import { getDetails } from "./src/fetch-details"
+
+export const renderDetails = async () => {
+    
+}
