+++ tasti-labs/src/fetch-category.js
@@ -4,6 +4,7 @@ export const getCategoryDishes = async (category) => {
         const response = await fetch(`https://themealdb.com/api/json/v1/1/filter.php?c=${category}`);
         if(!response.ok) throw new Error("Sorry can't fetch food!")
         const data = await response.json();
+    console.log(data)
         return data 
     }
     catch(error){
